Add id attribute value and disclaimer text
Added id value to a field that should change after users enter a dish
name, but wasn't changing when the dish wasn't found in the database.
Also added more text to the disclaimer area to further elaborate that
not having a 'bad' result doesn't necessarily mean the dish is safe.