Remove a console.log, reset content fields
Removed a console.log that was logging the return object from the api
fetch, as that was a development statement.
Added code to reset the text content field that displays a users entered
dish name.
Added code to display, even if not a dish, the dish name input by the
user, so they can see if they misspelt anything.